Move more junk into Board
This is clearly the start of a longer journey -- so much stuff in Board will need to go into its own child components
